% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{agregar_d_personas}
\alias{agregar_d_personas}
\title{Traspasa variables del conjunto D de la EU-SILC al conjunto P}
\usage{
agregar_d_personas(.P, .D)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.D}{\code{data.frame} o \code{tibble}. Conjunto de datos D de la EU-SILC.}
}
\value{
\code{tibble}. Conjunto P de la EU-SILC con algunas variables del conjunto D
}
\description{
Traspasa variables del conjunto D de la EU-SILC al conjunto P
}
