% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{agregar_r_personas}
\alias{agregar_r_personas}
\title{Traspasa variables del conjunto R de la EU-SILC al conjunto P}
\usage{
agregar_r_personas(.P, .R)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.R}{\code{data.frame} o \code{tibble}. Conjunto de datos R de la EU-SILC.}
}
\value{
\code{tibble}. Conjunto P de la EU-SILC con algunas varibles del conjunto R
}
\description{
Traspasa variables del conjunto R de la EU-SILC al conjunto P
}
