% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{agrupar_nac}
\alias{agrupar_nac}
\title{Agrupa los años de nacimiento en grupos de cinco años}
\usage{
agrupar_nac(.anio, .nac)
}
\arguments{
\item{.anio}{Año de la encuesta.}

\item{.nac}{Vector de años de nacimiento.}
}
\value{
Vector de años de nacimiento agrupados.
}
\description{
La función agrupa los años según el criterio que se aplica en el conjunto de
datos de Alemania. Los años de nacimiento se registran desde 81 años atrás
hasta el presente; aquellos que nacieron antes se agrupan en el primer año.
Los grupos de cinco años se arman a partir del primer año registrado. P.e.,
en 2023 el primer año registrado fue 2023 - 81 = 1942, por lo cual los
grupos resultan ser 1942-1946, 1947-1951, 1951-1956, etc.
}
