% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{aplicar_imputaciones}
\alias{aplicar_imputaciones}
\title{Title}
\usage{
aplicar_imputaciones(.datos, .anio, .lmh)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{.anio}

\item{.lmh}{.lmh}
}
\value{
Conjunto de datos estandarizado con imputaciones aplicadas
}
\description{
Title
}
