% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{armar_imputables}
\alias{armar_imputables}
\title{Title}
\usage{
armar_imputables(.datos, .imputadas, .predictoras, .flags, .factores = NULL)
}
\arguments{
\item{.datos}{.datos}

\item{.imputadas}{.imputadas}

\item{.predictoras}{.predictoras}

\item{.flags}{.flags}

\item{.factores}{.factores}
}
\value{
conjunto de datos para imputar
}
\description{
Title
}
