% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{aux_etiquetas}
\alias{aux_etiquetas}
\title{Title}
\usage{
aux_etiquetas(.etq)
}
\arguments{
\item{.etq}{df de xlsx de etiquetas}
}
\value{
lista anidada con etiquetas para etiquetar_eusilc
}
\description{
Title
}
