% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliares.R
\name{aux_imputaciones}
\alias{aux_imputaciones}
\title{Title}
\usage{
aux_imputaciones(.imp)
}
\arguments{
\item{.imp}{df de xlsx de imputaciones}
}
\value{
df de imputaciones para imputar_eusilc
}
\description{
Title
}
