% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_egp}
\alias{calc_egp}
\title{Calcula el clasificador de clase social de Erikson, Goldthorpe y Pontocarero}
\usage{
calc_egp(.PL051A, .PL040A, .PL150)
}
\arguments{
\item{.PL051A}{\code{numeric}. Ocupación (ISCO-08)}

\item{.PL040A}{\code{numeric}. Categoría ocupacional}

\item{.PL150}{\code{numeric}. Responsabilidades de supervisión}
}
\value{
\code{numeric}. Clasificador de clase de EGP
}
\description{
Calcula el clasificador de clase social de Erikson, Goldthorpe y Pontocarero
}
