% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{calc_flags_imputacion}
\alias{calc_flags_imputacion}
\title{Title}
\usage{
calc_flags_imputacion(.datos, .anio)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{.anio}
}
\value{
.datos con flags de imputacion
}
\description{
Title
}
