% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_informalidad}
\alias{calc_informalidad}
\title{Calcula el clasificador de informalidad laboral según aportes a la seguridad social}
\usage{
calc_informalidad(.PL040A, .PY030G, .PY035G, .nivel)
}
\arguments{
\item{.PL040A}{\code{numeric}. Categoría ocupacional}

\item{.PY030G}{\code{numeric}. Contribuciones a la seguridad social del empleador}

\item{.PY035G}{\code{numeric}. Contribuciones a pensiones privadas individuales}

\item{.nivel}{\code{numeric}. Nivel de agregación}
}
\value{
\code{numeric}. Clasificador de informalidad laboral
}
\description{
Calcula el clasificador de informalidad laboral según aportes a la seguridad social
}
