% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_personas.R
\name{calc_personas}
\alias{calc_personas}
\title{Construye variables en la base P de la EU-SILC.}
\usage{
calc_personas(.datos, .lmh = FALSE, ...)
}
\arguments{
\item{.datos}{Conjunto P de la EU-SILC.}

\item{.lmh}{Si el conjunto de datos tiene el modulo LMH.}

\item{...}{...}
}
\value{
Conjunto de datos P de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base P de la EU-SILC.
}
