% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_testablecimiento}
\alias{calc_testablecimiento}
\title{Title}
\usage{
calc_testablecimiento(.PL130, .nivel, .lmh = TRUE)
}
\arguments{
\item{.PL130}{PL130}

\item{.nivel}{Nivel de agregación.}

\item{.lmh}{Módulo LMH}
}
\value{
Tamaño del establecimiento
}
\description{
Title
}
