% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_variante_c}
\alias{calc_variante_c}
\title{Calcula híbrido entre características de la ocupación de los ocupados y los desocupados}
\usage{
calc_variante_c(.PL032, .a, .b)
}
\arguments{
\item{.PL032}{\code{numeric}. Condición de actividad}

\item{.a}{\code{numeric}. Características de la ocupación de los ocupados en el CRP}

\item{.b}{\code{numeric}. Características de la última ocupación de los desocupados en el CRP}
}
\value{
\code{numeric}. Características de la ocupación actual o de la última ocupación
}
\description{
Calcula híbrido entre características de la ocupación de los ocupados y los desocupados
}
