% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{calcular_auxiliares_personas}
\alias{calcular_auxiliares_personas}
\title{Calcula algunas variables auxiliares}
\usage{
calcular_auxiliares_personas(.P, .anio)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}
}
\value{
\code{tibble}. Conjunto P de la EU-SILC con variables auxiliares extra.
}
\description{
Calcula algunas variables auxiliares
}
