% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_hogares_panel.R
\name{calcular_hogares_panel_}
\alias{calcular_hogares_panel_}
\title{Construye variables del conjunto H longitudinal de la EU-SILC (interna)}
\usage{
calcular_hogares_panel_(.H)
}
\arguments{
\item{.H}{\code{data.frame} o \code{tibble}. Conjunto H longitudinal de la EU-SICL
estandarizado por \code{\link[=estandarizar_hogares_panel_]{estandarizar_hogares_panel_()}}.}
}
\value{
Conjunto H con las variables originales y las armonizadas.
}
\description{
Construye variables del conjunto H longitudinal de la EU-SILC (interna)
}
