% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calcular_personas_}
\alias{calcular_personas_}
\title{Title}
\usage{
calcular_personas_(.P)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC
estandarizado con \code{\link[=estandarizar_personas]{estandarizar_personas()}}.}
}
\value{
\code{tibble}. Conjunto de datos P de la EU-SILC estandarizado con variables armonizadas.
}
\description{
Title
}
