% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{chequear_bases_hogares}
\alias{chequear_bases_hogares}
\title{Chequea que los conjuntos H, P y D sean adecuados}
\usage{
chequear_bases_hogares(.H, .P, .D)
}
\arguments{
\item{.H}{Argumento .H}

\item{.P}{Argumento .P}

\item{.D}{Argumento .D}
}
\description{
Chequea que los conjuntos H, P y D sean adecuados
}
