% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_personas.R
\name{chequear_bases_personas}
\alias{chequear_bases_personas}
\title{Chequea que los conjuntos P, D y R sean adecuados}
\usage{
chequear_bases_personas(.P, .D, .R)
}
\arguments{
\item{.P}{Argumento .P}

\item{.D}{Argumento .D}

\item{.R}{Argumento .R}
}
\description{
Chequea que los conjuntos P, D y R sean adecuados
}
