% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{chequear_columnas}
\alias{chequear_columnas}
\title{Chequea la presencia de columnas requeridas}
\usage{
chequear_columnas(.datos, .columnas, .argumento = rlang::caller_arg(.datos))
}
\arguments{
\item{.datos}{Conjunto de datos.}

\item{.columnas}{Nombres de las columnas requeridas.}

\item{.argumento}{Nombre del argumento para el mensaje de error.}
}
\value{
\code{NULL}, invisiblemente.
}
\description{
Chequea la presencia de columnas requeridas
}
