% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{chequear_concordancia}
\alias{chequear_concordancia}
\title{Chequea la concordancia de una base auxiliar con la principal}
\usage{
chequear_concordancia(.referencia, .auxiliar, .principal, .argumento, .prefijo)
}
\arguments{
\item{.referencia}{Lista con anio y pais validos de la base principal.}

\item{.auxiliar}{Lista con anio y pais validos de la auxiliar.}

\item{.principal}{Nombre del argumento de la base principal.}

\item{.argumento}{Nombre del argumento de la base auxiliar.}

\item{.prefijo}{Prefijo de las clases de discrepancia: p, d o r.}
}
\value{
NULL, invisiblemente.
}
\description{
Chequea la concordancia de una base auxiliar con la principal
}
