% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{chequear_faltantes}
\alias{chequear_faltantes}
\title{Cuenta casos a imputar y casos de referencia}
\usage{
chequear_faltantes(.flag)
}
\arguments{
\item{.flag}{\code{numeric}. Vector de flags de imputación}
}
\value{
\code{character}. Cadena de caracteres que indica la decisión, "imputar", "no imputar" o "completo"
}
\description{
A partir de un flag de imputación cuenta los casos a imputar y los casos de
referencia y determina si se procede con la imputación, si no conviene por
la magnitud de la pérdida o si no hay nada que imputar.
}
