% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{chequear_insumos_perdidos}
\alias{chequear_insumos_perdidos}
\title{Chequea si alguno de los insumos esta completamente perdido}
\usage{
chequear_insumos_perdidos(...)
}
\arguments{
\item{...}{Vectores insumo}
}
\value{
\code{TRUE} si alguno de los insumos esta completamente perdido
}
\description{
Chequea si alguno de los insumos esta completamente perdido
}
