% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{chequear_perdidas}
\alias{chequear_perdidas}
\title{Chequea y avisa qué variables están completamente perdidas}
\usage{
chequear_perdidas(.datos, .base)
}
\arguments{
\item{.datos}{\code{tibble}. Conjunto de datos}

\item{.base}{\code{character}. Qué tipo de base es, P o H}
}
\value{
Nada
}
\description{
Chequea y avisa qué variables están completamente perdidas
}
