% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_variables.R
\name{construir_vbles_h}
\alias{construir_vbles_h}
\title{Construye variables en la base H de la EU-SILC.}
\usage{
construir_vbles_h(.datos, ...)
}
\arguments{
\item{.datos}{Conjunto H de la EU-SILC.}

\item{...}{...}
}
\value{
Conjunto H de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base H de la EU-SILC.
}
