% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_variables.R
\name{construir_vbles_p_lmh}
\alias{construir_vbles_p_lmh}
\title{Construye variables dependientes del módulo Labour Market and Housing de
la base P de la EU-SILC}
\usage{
construir_vbles_p_lmh(.datos)
}
\arguments{
\item{.datos}{Conjunto P de la EU-SILC expandido con \code{\link[=construir_vbles_p]{construir_vbles_p()}}.}
}
\value{
Conjunto P de la EU-SILC con variables adicionales.
}
\description{
Construye variables dependientes del módulo Labour Market and Housing de
la base P de la EU-SILC
}
