% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias_estandarizacion.R
\name{consultar_advertencias}
\alias{consultar_advertencias}
\title{Consulta las advertencias aplicables a un conjunto EU-SILC}
\usage{
consultar_advertencias(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto P o H, antes o despues de
procesarlo con el paquete. Debe corresponder a un unico anio y pais.}
}
\value{
\code{tibble}. Filas de \link{advertencias_estandarizacion} correspondientes
al anio, pais y base de \code{.datos}.
}
\description{
Consulta las advertencias aplicables a un conjunto EU-SILC
}
