% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{estandarizar_anio_personas}
\alias{estandarizar_anio_personas}
\title{Convierte algunas variables de los conjuntos previos a 2021 al formato que
tienen a partir de ese año}
\usage{
estandarizar_anio_personas(.P, .anio)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}
}
\value{
\code{tibble}. Conjunto de datos P con variables transformadas
}
\description{
Convierte algunas variables de los conjuntos previos a 2021 al formato que
tienen a partir de ese año
}
