% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_hogares.R
\name{estandarizar_hogares_}
\alias{estandarizar_hogares_}
\title{Estandariza el conjunto H de la EU-SILC para el proceso de armonización}
\usage{
estandarizar_hogares_(.H, .D, .anio, .pais)
}
\arguments{
\item{.H}{\code{data.frame}o \code{tibble}. Conjunto de datos H de la EU-SILC}

\item{.D}{\code{data.frame}o \code{tibble}. Conjunto de datos D de la EU-SILC}

\item{.anio}{\code{numeric}. Año al que corresponde el conjunto H}

\item{.pais}{\code{character}. País al que corresponde el conjunto H}
}
\value{
\code{tibble}. Conjunto de datos H de la EU-SILC estandarizado
}
\description{
Aplica transformaciones sobre las variables del conjunto H según el año, el
país y si se proveyó el conjunto D. El conjunto final tiene las variables
necesarias para aplicar \code{\link[=calcular_hogares]{calcular_hogares()}}. Las variables que no están
disponibles quedan como \code{NA}.
}
