% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{estandarizar_paises_personas}
\alias{estandarizar_paises_personas}
\title{Ajustes y advertencias específicas para ciertos países y años}
\usage{
estandarizar_paises_personas(.P, .anio, .pais)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}

\item{.pais}{\code{numeric}. País de la encuesta.}
}
\value{
\code{tibble}. Conjunto P de la EU-SILC con ajustes específicos por país
}
\description{
Ajustes y advertencias específicas para ciertos países y años
}
