% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{estandarizar_personas_}
\alias{estandarizar_personas_}
\title{Estandariza el conjunto P de la EU-SILC para el proceso de armonización (interna)}
\usage{
estandarizar_personas_(.P, .R, .D, .anio, .pais)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.R}{\code{data.frame} o \code{tibble}. Conjunto de datos R de la EU-SILC.}

\item{.D}{\code{data.frame} o \code{tibble}. Conjunto de datos D de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}

\item{.pais}{\code{character}. País de la encuesta.}
}
\value{
\code{tibble}. Conjunto de datos P estandarizado para \code{\link[=imputar_personas]{imputar_personas()}} y \code{\link[=calcular_personas]{calcular_personas()}}.
}
\description{
¡Esta función es interna! Aplica transformaciones sobre las variables del
conjunto P según el año, el país y si se proveyeron los conjuntos D y R. El
conjunto final tiene todas las variables necesarias para aplicar
\code{\link[=imputar_personas]{imputar_personas()}} y \code{\link[=calcular_personas]{calcular_personas()}}. Las variables que no están
disponibles quedan como \code{NA}.
}
\details{
Esta función es el núcleo interno de \code{\link[=estandarizar_personas]{estandarizar_personas()}}. Para más
detalles ver la documentación de esa función.
}
