% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetas.R
\docType{data}
\name{etiquetas}
\alias{etiquetas}
\title{Etiquetas de las variables y sus valores}
\format{
\subsection{\code{etiquetas}}{

A tibble: 336 x 6
\describe{
\item{variable}{Nombre de la variable en el conjunto de datos expandido}
\item{conjunto}{Conjunto de datos al que corresponde la variable (P o H)}
\item{bloque}{Bloque de variables al que corresponde}
\item{descripcion}{Descripcion de la variable}
\item{etiqueta}{Descripcion de la categoria}
\item{valor}{Codigo de la categoria}
}
}
}
\usage{
etiquetas
}
\description{
Contiene descripciones de cada variable y de las categorias de aquellas que
son factores.
}
\keyword{datasets}
