% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_eusilc.R
\name{expandir_eusilc}
\alias{expandir_eusilc}
\title{Title}
\usage{
expandir_eusilc(.datos, .base, ...)
}
\arguments{
\item{.datos}{descripcion}

\item{.base}{descripcion}

\item{...}{descripcion}
}
\value{
valor
}
\description{
Title
}
