% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_hogares.R
\name{f_agregar_personas}
\alias{f_agregar_personas}
\title{Agrega variables de ingreso de la base P de la EU-SILC a nivel hogar. (Optimizda)}
\usage{
f_agregar_personas(.personas)
}
\arguments{
\item{.personas}{Conjunto P de la EU-SILC expandido con \code{\link[=calcular_personas]{calcular_personas()}}.}
}
\value{
Conjunto de datos con ingresos individuales agregados a nivel hogar.
}
\description{
Agrega variables de ingreso de la base P de la EU-SILC a nivel hogar. (Optimizda)
}
