% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_horas}
\alias{imputar_horas}
\title{Imputa horas semanales habitualmente trabajadas}
\usage{
imputar_horas(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa horas semanales habitualmente trabajadas
}
