% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_laboral_a}
\alias{imputar_laboral_a}
\title{Imputa categoría ocupacional, ocupación y rama de actividad (A)}
\usage{
imputar_laboral_a(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa categoría ocupacional, ocupación y rama de actividad (A)
}
