% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_laboral_b}
\alias{imputar_laboral_b}
\title{Imputa categoría ocupacional, ocupación y rama de actividad (B)}
\usage{
imputar_laboral_b(.datos, .anio)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}

\item{.anio}{\code{numeric}. El año al que corresponde el conjunto de datos}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa categoría ocupacional, ocupación y rama de actividad (B)
}
