% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_meses}
\alias{imputar_meses}
\title{Imputa meses con ingresos en el IRP}
\usage{
imputar_meses(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa meses con ingresos en el IRP
}
