% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_personas_}
\alias{imputar_personas_}
\title{Title}
\usage{
imputar_personas_(.P, .anio, .pais)
}
\arguments{
\item{.P}{\code{data.frame} o \code{tibble}. Conjunto de datos P de la EU-SILC
estandarizado con \code{\link[=estandarizar_personas]{estandarizar_personas()}}.}

\item{.anio}{\code{numeric}. Año de la encuesta.}

\item{.pais}{\code{character}. País de la encuesta.}
}
\value{
\code{tibble}. Conjunto P de la EU-SILC con valores imputados.
}
\description{
Title
}
