% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_sectorpp}
\alias{imputar_sectorpp}
\title{Imputa el sector de actividad (público o privado)}
\usage{
imputar_sectorpp(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa el sector de actividad (público o privado)
}
