% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_tamanio}
\alias{imputar_tamanio}
\title{Imputa el tamaño del establecimiento}
\usage{
imputar_tamanio(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P a imputar}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Imputa el tamaño del establecimiento
}
