% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias.R
\name{informar_disponibilidad_modulos}
\alias{informar_disponibilidad_modulos}
\title{Advierte sobre la disponibilidad y ausencia de las variables PL130 y PL230}
\usage{
informar_disponibilidad_modulos(.P, .anio)
}
\arguments{
\item{.P}{\code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}
}
\value{
Nada.
}
\description{
Advierte sobre la disponibilidad y ausencia de las variables PL130 y PL230
}
