% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias.R
\name{informar_insumos_hogares}
\alias{informar_insumos_hogares}
\title{Advierte la pérdida de variables ante la falta de insumos}
\usage{
informar_insumos_hogares(.D)
}
\arguments{
\item{.D}{\code{tibble}. Conjunto de datos D de la EU-SILC.}
}
\value{
Nada.
}
\description{
Advierte la pérdida de variables ante la falta de insumos
}
