% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias.R
\name{informar_insumos_personas}
\alias{informar_insumos_personas}
\title{Advierte la pérdida de variables ante la falta de insumos}
\usage{
informar_insumos_personas(.P, .D, .R, .anio)
}
\arguments{
\item{.P}{\code{tibble}. Conjunto de datos P de la EU-SILC.}

\item{.D}{\code{tibble}. Conjunto de datos D de la EU-SILC.}

\item{.R}{\code{tibble}. Conjunto de datos R de la EU-SILC.}

\item{.anio}{\code{numeric}. Año de la encuesta.}
}
\value{
Nada.
}
\description{
Advierte la pérdida de variables ante la falta de insumos
}
