% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{informar_pais_no_probado}
\alias{informar_pais_no_probado}
\title{Informa si el pais validado no fue probado con el paquete}
\usage{
informar_pais_no_probado(.periodo)
}
\arguments{
\item{.periodo}{Lista con anio y pais validos.}
}
\value{
NULL, invisiblemente.
}
\description{
Informa si el pais validado no fue probado con el paquete
}
