% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{lookup_personas}
\alias{lookup_personas}
\title{Title}
\usage{
lookup_personas(.datos, .anio, .lmh = FALSE, ...)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{anio}

\item{.lmh}{.lmh}

\item{...}{...}
}
\value{
.datos con varibles recodificadas
}
\description{
Title
}
