% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias.R
\name{obtener_advertencias}
\alias{obtener_advertencias}
\title{Obtiene las advertencias correspondientes a un país, año y tipo de base}
\usage{
obtener_advertencias(.base, .anio, .pais)
}
\arguments{
\item{.base}{\code{character}. Tipo de base, 'P' o 'H'.}

\item{.anio}{\code{numeric}. Año de la encuesta.}

\item{.pais}{\code{character}. País de la encuesta.}
}
\value{
\code{tibble}. Advertencias por variable para el país, año y base.
}
\description{
Obtiene las advertencias correspondientes a un país, año y tipo de base
}
