% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chequear.R
\name{obtener_periodo_base}
\alias{obtener_periodo_base}
\title{Obtiene el unico anio y pais validos de una base}
\usage{
obtener_periodo_base(.datos, .columnas, .argumento = rlang::caller_arg(.datos))
}
\arguments{
\item{.datos}{Data frame de una base suministrada.}

\item{.columnas}{Nombres de las columnas de anio y pais, en ese orden.}

\item{.argumento}{Nombre del argumento para los mensajes.}
}
\value{
Lista con anio y pais.
}
\description{
Obtiene el unico anio y pais validos de una base
}
