% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsa.eusilc-package.R
\docType{package}
\name{odsa.eusilc-package}
\alias{odsa.eusilc}
\alias{odsa.eusilc-package}
\title{odsa.eusilc: Funciones para trabajar con los datos de la encuesta EU-SILC}
\description{
Se proporcionan funciones para construir nuevas variables, imputar datos faltantes y asignar etiquetas a los conjuntos de datos de la encuesta EU-SILC de acuerdo con el uso habitual que se les da en el ODSA.
}
\author{
\strong{Maintainer}: Fernando Piderit \email{fgpiderit@gmail.com}

Authors:
\itemize{
  \item Fernando Piderit \email{fgpiderit@gmail.com}
}

}
\keyword{internal}
