% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advertencias.R
\name{ver_advertencias}
\alias{ver_advertencias}
\title{Muestra las advertencias correspondientes al país, año y tipo de base suministrada}
\usage{
ver_advertencias(.datos)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos EU-SILC expandido o estandarizado.}
}
\value{
\code{tibble}. Conjunto de datos con información sobre las advertencias.
}
\description{
Muestra las advertencias correspondientes al país, año y tipo de base suministrada
}
