% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_variables.R
\name{agregar_personas}
\alias{agregar_personas}
\title{Agrega variables de ingreso de la base P de la EU-SILC a nivel hogar.}
\usage{
agregar_personas(.personas)
}
\arguments{
\item{.personas}{Conjunto P de la EU-SILC expandido con \code{\link[=construir_vbles_p]{construir_vbles_p()}} y, opcionalmente, con \code{\link[=construir_vbles_p_lmh]{construir_vbles_p_lmh()}}.}
}
\value{
Conjunto de datos con ingresos individuales agregados a nivel hogar.
}
\description{
Agrega variables de ingreso de la base P de la EU-SILC a nivel hogar.
}
