% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{aplicar_imputaciones}
\alias{aplicar_imputaciones}
\title{Title}
\usage{
aplicar_imputaciones(.datos, .imp, .vble, .flag)
}
\arguments{
\item{.datos}{.datos}

\item{.imp}{.imp}

\item{.vble}{.vble}

\item{.flag}{.flag}
}
\value{
Conjunto de datos con las imputaciones aplicadas
}
\description{
Title
}
