% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{aplicar_imputaciones}
\alias{aplicar_imputaciones}
\title{Reemplaza valores originales por valores imputados}
\usage{
aplicar_imputaciones(.datos, .imp, .vble, .flag)
}
\arguments{
\item{.datos}{\code{data.frame} o \code{tibble}. Conjunto de datos P sin imputar}

\item{.imp}{\code{data.frame} o \code{tibble}. Conjunto de datos con las observaciones imputadas}

\item{.vble}{\code{character}. Nombre de la variable imputada}

\item{.flag}{\code{character}. Nombre del flag de imputación de la variable imputada}
}
\value{
\code{tibble}. Conjunto de datos P con valores imputados
}
\description{
Dado un conjunto de datos P y un conjunto de datos con imputaciones,
reemplaza los valores de la variable original por los valores imputados
para los casos corresopondientes.
}
