% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_personas.R
\name{aplicar_imputaciones}
\alias{aplicar_imputaciones}
\title{Title}
\usage{
aplicar_imputaciones(.datos)
}
\arguments{
\item{.datos}{.datos}
}
\value{
Conjunto de datos estandarizado con imputaciones aplicadas
}
\description{
Title
}
