% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetar_eusilc.R
\name{armar_etiquetas}
\alias{armar_etiquetas}
\title{Title}
\usage{
armar_etiquetas(.etq)
}
\arguments{
\item{.etq}{df de xlsx de etiquetas}
}
\value{
lista anidada con etiquetas para etiquetar_eusilc
}
\description{
Title
}
