% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_calidad}
\alias{calc_calidad}
\title{Calcula el indicador de empleo de calidad}
\usage{
calc_calidad(.pl40a, .pl12a, .pomj)
}
\arguments{
\item{.pl40a}{\code{numeric}. Informalidad y asalarización}

\item{.pl12a}{\code{numeric}. Calificación de la ocupación}

\item{.pomj}{\code{numeric}. Permanencia del trabajo principal}
}
\value{
Indicador de empleo de calidad
}
\description{
Calcula el indicador de empleo de calidad
}
