% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_egp}
\alias{calc_egp}
\title{Title}
\usage{
calc_egp(.PL051A, .PL040A, .PL150)
}
\arguments{
\item{.PL051A}{PL051A}

\item{.PL040A}{PL040A}

\item{.PL150}{PL150}
}
\value{
egp
}
\description{
Title
}
