% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_personas.R
\name{calc_heterogeneidad}
\alias{calc_heterogeneidad}
\title{Title}
\usage{
calc_heterogeneidad(.PL040A, .pl06b, .pl08b, .lmh = TRUE)
}
\arguments{
\item{.PL040A}{PL040A}

\item{.pl06b}{pl06b}

\item{.pl08b}{pl08b}

\item{.lmh}{lmh}
}
\value{
heterogeneidad
}
\description{
Title
}
