% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_heterogeneidad}
\alias{calc_heterogeneidad}
\title{Title}
\usage{
calc_heterogeneidad(
  .PL040A,
  .PL032,
  .pl05,
  .pl06b,
  .pl07,
  .pl08b,
  .nivel,
  .lmh = TRUE
)
}
\arguments{
\item{.PL040A}{PL040A}

\item{.PL032}{PL032}

\item{.pl05}{pl05}

\item{.pl06b}{pl06b}

\item{.pl07}{pl07}

\item{.pl08b}{pl08b}

\item{.nivel}{Nivel de agregación.}

\item{.lmh}{lmh}
}
\value{
heterogeneidad
}
\description{
Title
}
