% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_heterogeneidad}
\alias{calc_heterogeneidad}
\title{Title}
\usage{
calc_heterogeneidad(
  .PL040A,
  .PL032,
  .pl20,
  .pl21b,
  .pl22,
  .pl13,
  .nivel,
  .lmh = TRUE
)
}
\arguments{
\item{.PL040A}{PL040A}

\item{.PL032}{PL032}

\item{.pl20}{pl20}

\item{.pl21b}{pl21b}

\item{.pl22}{pl22}

\item{.pl13}{pl13}

\item{.nivel}{Nivel de agregación.}

\item{.lmh}{lmh}
}
\value{
heterogeneidad
}
\description{
Title
}
