% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_informalidad}
\alias{calc_informalidad}
\title{Title}
\usage{
calc_informalidad(.PL040A, .PY030G, .PY035G, .nivel)
}
\arguments{
\item{.PL040A}{PL040A}

\item{.PY030G}{PY030G}

\item{.PY035G}{PY035G}

\item{.nivel}{nivel de agregación}
}
\value{
informalidad
}
\description{
Title
}
