% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_y_sector}
\alias{calc_y_sector}
\title{Censura ingresos que no provienen de determinado sector de inserción}
\usage{
calc_y_sector(.py10, .pl31, .sector)
}
\arguments{
\item{.py10}{\code{numeric}. Ingresos laborales}

\item{.pl31}{\code{numeric}. Sector de inserción de los individuos}

\item{.sector}{\code{numeric}. Sector de inserción a seleccionar}
}
\value{
\code{numeric}. Ingresos laborales provenientes de determinado sector, 0 el resto
}
\description{
Censura ingresos que no provienen de determinado sector de inserción
}
