% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calc_y_sector}
\alias{calc_y_sector}
\title{Title}
\usage{
calc_y_sector(.py10, .pl31, .sector)
}
\arguments{
\item{.py10}{py10}

\item{.pl31}{pl31}

\item{.sector}{sector}
}
\value{
py13, py14 o py15
}
\description{
Title
}
