% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_hogares.R
\name{calcular_hogares}
\alias{calcular_hogares}
\title{Construye variables en la base H de la EU-SILC.}
\usage{
calcular_hogares(.datos, ...)
}
\arguments{
\item{.datos}{Conjunto H de la EU-SILC.}

\item{...}{...}
}
\value{
Conjunto H de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base H de la EU-SILC.
}
