% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_hogares.R
\name{calcular_hogares_}
\alias{calcular_hogares_}
\title{Construye variables en la base H de la EU-SILC.}
\usage{
calcular_hogares_(.H)
}
\arguments{
\item{.H}{Conjunto H de la EU-SILC.}
}
\value{
Conjunto H de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base H de la EU-SILC.
}
