% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_personas.R
\name{calcular_personas}
\alias{calcular_personas}
\title{Construye variables en la base P de la EU-SILC.}
\usage{
calcular_personas(.datos, .anio, .lmh, ...)
}
\arguments{
\item{.datos}{Conjunto P de la EU-SILC.}

\item{.anio}{El año del conjunto}

\item{.lmh}{Si el conjunto de datos tiene el modulo LMH.}

\item{...}{...}
}
\value{
Conjunto de datos P de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base P de la EU-SILC.
}
