% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construir_variables.R
\name{construir_vbles_p}
\alias{construir_vbles_p}
\title{Construye variables en la base P de la EU-SILC.}
\usage{
construir_vbles_p(.datos, ...)
}
\arguments{
\item{.datos}{Conjunto P de la EU-SILC.}

\item{...}{...}
}
\value{
Conjunto de datos P de la EU-SILC con variables adicionales.
}
\description{
Construye variables en la base P de la EU-SILC.
}
