% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_hogares.R
\name{estandarizar_hogares}
\alias{estandarizar_hogares}
\title{Title}
\usage{
estandarizar_hogares(.datos, .anio, .pais, .D, .lmh)
}
\arguments{
\item{.datos}{Conjunto de datos H}

\item{.anio}{Año de la encuesta}

\item{.pais}{Pais de la encuesta}

\item{.D}{Conjunto de datos D}

\item{.lmh}{Si se encontró la variable PL230}
}
\value{
Conjunto de datos H estandarizado
}
\description{
Title
}
