% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estandarizar_personas.R
\name{estandarizar_personas}
\alias{estandarizar_personas}
\title{Title}
\usage{
estandarizar_personas(.datos, .anio, .D, .R, .lmh)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{.anio}

\item{.D}{.D}

\item{.R}{.R}

\item{.lmh}{lmh}
}
\value{
conjunto de datos estandarizado para \code{\link[=calcular_personas]{calcular_personas()}}.
}
\description{
Title
}
