% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_personas.R
\name{estandarizar_personas}
\alias{estandarizar_personas}
\title{Title}
\usage{
estandarizar_personas(.datos, .anio, .bloques, .D, .R)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{.anio}

\item{.bloques}{.bloques}

\item{.D}{.D}

\item{.R}{.R}
}
\value{
conjunto de datos estandarizado para \code{\link[=calc_personas]{calc_personas()}}.
}
\description{
Title
}
