% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetar_eusilc.R
\name{etiquetar_eusilc}
\alias{etiquetar_eusilc}
\title{Title}
\usage{
etiquetar_eusilc(.datos, ...)
}
\arguments{
\item{.datos}{datos}

\item{...}{...}
}
\value{
datos etiquetados
}
\description{
Title
}
