% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etiquetar_eusilc.R
\name{etiquetar_eusilc_}
\alias{etiquetar_eusilc_}
\title{Title}
\usage{
etiquetar_eusilc_(.datos, .base)
}
\arguments{
\item{.datos}{datos}

\item{.base}{base}
}
\value{
datos etiquetados
}
\description{
Title
}
