% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_eusilc.R
\name{expandir_eusilc}
\alias{expandir_eusilc}
\title{Construir variables adicionales en los conjuntos de datos de la EU-SILC}
\usage{
expandir_eusilc(
  .datos,
  .base,
  .pais,
  ...,
  .individuos = NULL,
  .mantener = FALSE
)
}
\arguments{
\item{.datos}{Conjunto de datos de la EU-SILC. Debe ser el conjunto P o el conjunto H.}

\item{.base}{El tipo de conjunto de datos. Debe ser P o H.}

\item{.pais}{El pais al que corresponde el conjunto de datos.}

\item{...}{...}

\item{.individuos}{Conjunto de datos P expandido por \code{expandir_eusilc()}. Sólo si \code{.base} es H.}

\item{.mantener}{Conservar las variables originales en el conjunto de datos final o eliminarlas.}
}
\value{
Conjunto de datos de la EU-SILC con variables adicionales de uso habitual.
}
\description{
Construir variables adicionales en los conjuntos de datos de la EU-SILC
}
