% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandir_personas.R
\name{expandir_personas}
\alias{expandir_personas}
\title{Construir variables adicionales en los conjuntos de datos P de la EU-SILC}
\usage{
expandir_personas(.datos, ..., .D = NULL, .R = NULL, .mantener = FALSE)
}
\arguments{
\item{.datos}{Conjunto de datos P de la EU-SILC.}

\item{...}{...}

\item{.D}{Conjunto de datos D de la EU-SILC.}

\item{.R}{Conjunto de datos R de la EU-SILC.}

\item{.mantener}{Conservar las variables originales en el conjunto de datos final o eliminarlas.}
}
\value{
Conjunto de datos de la EU-SILC con variables adicionales de uso habitual.
}
\description{
Construir variables adicionales en los conjuntos de datos P de la EU-SILC
}
