% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_laboral_b}
\alias{imputar_laboral_b}
\title{Title}
\usage{
imputar_laboral_b(.datos, .anio)
}
\arguments{
\item{.datos}{.datos}

\item{.anio}{.anio}
}
\value{
datos imputados
}
\description{
Title
}
