% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputar_personas.R
\name{imputar_personas}
\alias{imputar_personas}
\title{Title}
\usage{
imputar_personas(.datos, .variables, ...)
}
\arguments{
\item{.datos}{Conjunto de datos P a imputar.}

\item{.variables}{Grupo de variables a imputar.}

\item{...}{...}
}
\value{
Conjunto de datos P con variables imputadas.
}
\description{
Title
}
