% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_ppa.R
\docType{data}
\name{tabla_ppa}
\alias{tabla_ppa}
\title{Tabla de factores de conversión a unidades de Paridad de Poder Adquisitivo}
\format{
tabla_ppa
A tibble: 390 x 3
\describe{
\item{PB010}{Año}
\item{PB020}{País}
\item{ppa_factor}{Factor de conversión a unidades de PPA de la Unión Europea 27 países}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/purchasing-power-parities/database}
}
\usage{
tabla_ppa
}
\description{
Indica el valor del factor de conversión de unidades de la moneda nacional a
unidades de paridad de poder adquisitivo (PPA) relativo a la Unión Europea 27
países.
}
\details{
La tabla se utiliza para convertir los montos de ingreso de los diferentes
países a unidades de PPA en dólares. \code{\link[=calcular_personas]{calcular_personas()}} y
\code{\link[=calcular_hogares]{calcular_hogares()}} utilizan las variables PX010 y HX010 (\emph{change rate euro
to national currency}) para convertir los montos a moneda nacional,
\code{ppa_factor} para convertirlos a unidades de PPA en euros de la Unión
Europea 27 países y finalmente la inversa del factor de conversión de los
Estados Unidos para convertirlos a unidades de PPA en dólares estadounidenses.
}
\keyword{datasets}
