% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_ppa.R
\docType{data}
\name{tabla_ppa}
\alias{tabla_ppa}
\title{Tabla de factores de conversión a unidades de Paridad de Poder Adquisitivo}
\format{
tabla_ppa
A tibble: 56 x 3
\describe{
\item{PB010}{Año}
\item{PB020}{País}
\item{ppa}{Factor de conversión a unidades de PPA}
}
}
\source{
\url{https://ec.europa.eu/eurostat/web/purchasing-power-parities/database}
}
\usage{
tabla_ppa
}
\description{
Indica el valor del factor de conversión de unidades de la moneda nacional
a unidades de paridad de poder adquisitivo relativo a la Unión Europea 27
países en 2020.
}
\keyword{datasets}
